+++ packages/atomic/src/components/atomic-result/atomic-result.tsx
@@ -1,5 +1,5 @@
 import {Component, h, Prop, Element, Listen} from '@stencil/core';
-import {Result, Engine} from '@coveo/headless';
+import {Result, SearchEngine} from '@coveo/headless';
 import {bindLogDocumentOpenOnResult} from '../../utils/result-utils';
 
 /**
@@ -19,9 +19,9 @@ export class AtomicResult {
   @Prop() result!: Result;
 
   /**
-   * The Headless engine.
+   * The headless search engine.
    */
-  @Prop() engine!: Engine;
+  @Prop() engine!: SearchEngine;
 
   /**
    * The result content to display.