+++ packages/headless/src/app/product-recommendation-engine/product-recommendation-engine.ts
@@ -16,7 +16,7 @@ import {
   ExternalEngineOptions,
 } from '../engine';
 import {buildLogger} from '../logger';
-import {searchHub} from '../reducers';
+import {productRecommendations, searchHub} from '../reducers';
 import {buildThunkExtraArguments} from '../thunk-extra-arguments';
 import {
   ProductRecommendationEngineConfiguration,
@@ -24,7 +24,12 @@ import {
 } from './product-recommendation-engine-configuration';
 import {SearchThunkExtraArguments} from '../headless-engine';
 
-const productRecommendationEngineReducers = {searchHub};
+export {
+  ProductRecommendationEngineConfiguration,
+  getSampleProductRecommendationEngineConfiguration,
+} from './product-recommendation-engine-configuration';
+
+const productRecommendationEngineReducers = {searchHub, productRecommendations};
 type ProductRecommendationEngineReducers = typeof productRecommendationEngineReducers;
 type ProductRecommendationEngineState = StateFromReducersMapObject<
   ProductRecommendationEngineReducers
@@ -41,7 +46,7 @@ export interface ProductRecommendationEngine
   > {}
 
 /**
- * The product recommendation engine configuration options.
+ * The product recommendation engine options.
  */
 export interface ProductRecommendationEngineOptions
   extends ExternalEngineOptions<ProductRecommendationEngineState> {