+++ packages/headless/src/test/mock-search-response.ts
@@ -3,6 +3,7 @@ import {
   SearchResponseSuccessWithDebugInfo,
 } from '../api/search/search/search-response';
 import {emptyCorrection} from '../features/did-you-mean/did-you-mean-state';
+import {emptyQuestionAnswer} from '../features/search/search-state';
 
 export function buildMockSearchResponse(
   config: Partial<SearchResponseSuccess> = {}
@@ -14,6 +15,7 @@ export function buildMockSearchResponse(
     facets: [],
     queryCorrections: [emptyCorrection()],
     triggers: [],
+    questionAnswer: emptyQuestionAnswer(),
     ...config,
   };
 }