+++ packages/quantic/force-app/main/default/lwc/quanticHeadlessLoader/__tests__/headlessLoader.test.js
@@ -6,11 +6,10 @@ import {
   getHeadlessEngine,
   initializeWithHeadless
 } from '../quanticHeadlessLoader';
-import { CoveoHeadlessStub, MockEngine } from '../../testUtils/coveoHeadlessStub';
+import { CoveoHeadlessStub } from '../../testUtils/coveoHeadlessStub';
 import {Deferred} from '../../quanticUtils/quanticUtils';
 
 describe('c/quanticHeadlessLoader', () => {
-  const dispatchMock = jest.fn();
   const testConfig = {
     organizationId: 'testOrganization',
     accessToken: 'bogus-token-xxxxx-xxxxx',
@@ -122,9 +121,7 @@ describe('c/quanticHeadlessLoader', () => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
-            engine: {
-              dispatch: dispatchMock
-            }
+            engine: CoveoHeadlessStub.buildSearchEngine()
           }
         }
       });
@@ -164,9 +161,7 @@ describe('c/quanticHeadlessLoader', () => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
-            engine: Promise.resolve({
-              dispatch: dispatchMock
-            }),
+            engine: Promise.resolve(CoveoHeadlessStub.buildSearchEngine()),
             config: new Deferred()
           }
         }
@@ -183,12 +178,13 @@ describe('c/quanticHeadlessLoader', () => {
           ];
         });
 
-        it('should set the component to initialized', () => {
+        it('should set the component to initialized', async () => {
           setComponentInitialized(testElement, testId);
   
           assertComponentIsSetInitialized(testElement, testId);
           jest.runAllTimers();
-          expect(dispatchMock).not.toBeCalled();
+          const engine = await window.coveoHeadless[testId].engine;
+          expect(engine.executeFirstSearch).not.toBeCalled();
         });
       });
 
@@ -209,7 +205,8 @@ describe('c/quanticHeadlessLoader', () => {
           assertComponentIsSetInitialized(testElement, testId);
           jest.runAllTimers();
           await window.coveoHeadless.engine;
-          expect(dispatchMock).toBeCalled();
+          const engine = await window.coveoHeadless[testId].engine;
+          expect(engine.executeFirstSearch).toBeCalled();
         });
       });
     });
@@ -229,7 +226,7 @@ describe('c/quanticHeadlessLoader', () => {
       it('should init the engine return a promise that resolves to that instance', async () => {
         const engine = await getHeadlessEngine(testId);
 
-        expect(engine).toBeInstanceOf(MockEngine);
+        expect(engine).toBeTruthy();
       });
     });
 
@@ -294,7 +291,7 @@ describe('c/quanticHeadlessLoader', () => {
         await initializeWithHeadless(testElement, testId, initialize);
         const engine = await window.coveoHeadless[testId].engine;
 
-        expect(engine).toBeInstanceOf(MockEngine);
+        expect(engine).toBeTruthy();
         expect(initialize).toHaveBeenCalled();
         assertComponentIsSetInitialized(testElement, testId);
       });