+++ packages/samples/headless-react/src/pages/SearchPage.tsx
@@ -49,13 +49,9 @@ import {StandaloneSearchBox as StandaloneSearchBoxFn} from '../components/standa
 import {RedirectionTrigger} from '../components/triggers/redirection-trigger.class';
 import {RedirectionTrigger as RedirectionTriggerFn} from '../components/triggers/redirection-trigger.fn';
 import {
-  HeadlessEngine,
-  Engine,
-  searchAppReducers,
-  recommendationAppReducers,
-  RecommendationAppState,
-  SearchActions,
-  AnalyticsActions,
+  SearchEngine,
+  buildSearchEngine,
+  getSampleSearchEngineConfiguration,
   Unsubscribe,
   RecommendationList as HeadlessRecommendationList,
   buildRecommendationList,
@@ -108,6 +104,11 @@ import {
   RedirectionTrigger as HeadlessRedirectionTrigger,
   buildRedirectionTrigger,
 } from '@coveo/headless';
+import {
+  RecommendationEngine,
+  buildRecommendationEngine,
+  getSampleRecommendationEngineConfiguration,
+} from '@coveo/headless/recommendation';
 import {bindUrlManager} from '../components/url-manager/url-manager';
 import {setContext} from '../components/context/context';
 import {dateRanges} from '../components/date-facet/date-utils';
@@ -126,8 +127,8 @@ const initialCriterion = criteria[0][1];
 const resultsPerPageOptions = [10, 25, 50, 100];
 
 export class SearchPage extends Component {
-  private readonly engine: Engine;
-  private readonly recommendationEngine: Engine<RecommendationAppState>;
+  private readonly engine: SearchEngine;
+  private readonly recommendationEngine: RecommendationEngine;
 
   private readonly recommendationList: HeadlessRecommendationList;
   private readonly tabs: {
@@ -163,14 +164,12 @@ export class SearchPage extends Component {
   constructor(props: {}) {
     super(props);
 
-    this.engine = new HeadlessEngine({
-      configuration: HeadlessEngine.getSampleConfiguration(),
-      reducers: searchAppReducers,
+    this.engine = buildSearchEngine({
+      configuration: getSampleSearchEngineConfiguration(),
     });
 
-    this.recommendationEngine = new HeadlessEngine({
-      configuration: HeadlessEngine.getSampleConfiguration(),
-      reducers: recommendationAppReducers,
+    this.recommendationEngine = buildRecommendationEngine({
+      configuration: getSampleRecommendationEngineConfiguration(),
     });
 
     this.recommendationList = buildRecommendationList(
@@ -297,9 +296,7 @@ export class SearchPage extends Component {
   }
 
   private executeInitialSearch() {
-    this.engine.dispatch(
-      SearchActions.executeSearch(AnalyticsActions.logInterfaceLoad())
-    );
+    this.engine.executeFirstSearch();
   }
 
   private updateAnalyticsContext() {